+++ CHANGELOG.md
@@ -11,7 +11,7 @@ All notable changes to `:vips` will be documented in this file.
 - type annotations [Kleis Auke Wolthuizen]
 - libvips draw calls now work
 - logging, see Vips\Image::setLogging()
-- exceptions on error
+- throw Vips\Exception on error
 
 ### Deprecated
 - now require php >= 7.0.11, fixes #10